restructure expander so that exapnd string is a standalone function
