+++ src/expander/expander_utils.c
@@ -1,38 +1,68 @@
 #include "shell.h"
 
-char	*remove_whitespaces(char *str);
-
-char	*append_str(char *str, char *to_append, size_t i, size_t j)
+char	*remove_whitespaces(char *str)
 {
-	char	*append;
+	size_t	i;
+	size_t	j;
+	char	*res;
 
-	if (to_append == NULL)
-		return (ft_substr(str, j, i - j));
-	append = ft_substr(str, j, i - j);
-	to_append = ft_strjoin_f(to_append, append);
-	free(append);
-	return (to_append);
+	i = 0;
+	j = 0;
+	if (str == NULL || *str == 0)
+		return (NULL);
+	res = ft_calloc(ft_strlen(str), sizeof(char *));
+	while (str[i])
+	{
+		if (str[i] == 9)
+			res[j++] = ' ';
+		else
+			res[j++] = str[i];
+		if (str[i] == 32 || str[i] == 9)
+			while (str[i] == 32 || str[i] == 9)
+				i++;
+		else
+			i++;
+	}
+	free(str);
+	str = res;
+	res = ft_strdup(str);
+	return (free(str), res);
 }
 
-char	*append_value(char *to_append, char *name, char **env)
+char	*trim_front(char *s1, const char *set)
 {
-	char	*value;
+	char	*res;
+	int		start;
+	int		len;
 
-	if (!ft_strncmp(name, "?", 1))
-	{
-		value = ft_itoa(g_exit_code);
-		if (ft_strlen(name) > 1)
-			value = ft_strjoin_f(value, name + 1);
-		free(name);
-	}
-	else
-		value = remove_whitespaces(get_env_value(name, env));
-	if (value == NULL)
-		return (to_append);
-	if (to_append == NULL || *to_append == 0)
-		return (trim_front(value, " \t\n\v"));
-	to_append = ft_strjoin_f(to_append, value);
-	return (free(value), to_append);
+	if (!s1 || !set)
+		return (NULL);
+	start = 0;
+	len = ft_strlen(s1);
+	while (s1[start] && ft_strchr(set, s1[start]))
+		start++;
+	res = ft_calloc((len - start + 1), sizeof(*s1));
+	if (!res)
+		return (NULL);
+	ft_strlcpy(res, s1 + start, len - start + 1);
+	return (free(s1), res);
+}
+
+char	*trim_back(char *s1, const char *set)
+{
+	char	*res;
+	int		end;
+
+	if (!s1 || !set)
+		return (NULL);
+	end = ft_strlen(s1);
+	while (end > 0 && ft_strchr(set, s1[end - 1]))
+		end--;
+	res = ft_calloc((end + 1), sizeof(*s1));
+	if (!res)
+		return (NULL);
+	ft_strlcpy(res, s1, end + 1);
+	return (free(s1), res);
 }
 
 char	*get_env_value(char *varname, char **env)
